+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.0] - 20 -07-2024
+
+### Updated
+
+- Product Reviews
+  - [x] Enabled users to write reviews for products, rate them and provide feedback.
+- Ensures to update views in [product detail component](./frontend/e-commerce-service/src/app/components/product-management/product-detail/) for any user
+
 ## [0.9.0] - 19-07-2024
 
 ### Updated